requirements.txt
asgiref==3.5.0
Django==3.2.13
pip==22.0.4
setuptools==60.10.0
sqlparse==0.4.2
wheel==0.37.1
python-dotenv==0.19.2
psycopg2==2.9.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
django-rest-swagger==2.2.0
django-celery-beat==2.2.1
django-celery-results==2.3.0
redis==4.2.2
pyyaml==6.0
